Create `useCachedWalletName` hook
