+++ src/shared/constants/local-storage.ts
@@ -5,3 +5,4 @@ export const LOCAL_STORAGE_BRAVE = "taho.brave"
 export const LOCAL_STORAGE_COOKIES = "taho.cookies"
 export const LOCAL_STORAGE_DISPLAYED_REALMS = "taho.displayedRealm"
 export const LOCAL_STORAGE_DISPLAYED_CHALLENGES = "taho.displayedChallenges"
+export const LOCAL_STORAGE_CACHED_NAMES = "taho.cachedNames"
